fix(replays): fix placement of mobile banner
