+++ static/app/components/replays/alerts/replayUnsupportedAlert.tsx
@@ -1,3 +1,4 @@
+import styled from '@emotion/styled';
 import Alert from 'sentry/components/alert';
 import ExternalLink from 'sentry/components/links/externalLink';
 import {IconInfo} from 'sentry/icons';
@@ -12,11 +13,15 @@ export default function ReplayUnsupportedAlert({projectSlug}: Props) {
     <ExternalLink href="https://docs.sentry.io/product/session-replay/getting-started/#supported-sdks" />
   );
   return (
-    <Alert icon={<IconInfo />}>
+    <StyledAlert icon={<IconInfo />}>
       <strong>{t(`Session Replay isn't available for %s.`, projectSlug)}</strong>{' '}
       {tct(`[docsLink: See our docs] to find out which platforms are supported.`, {
         docsLink,
       })}
-    </Alert>
+    </StyledAlert>
   );
 }
+
+const StyledAlert = styled(Alert)`
+  margin: 0;
+`;
