+++ static/app/views/performance/transactionSummary/transactionReplays/transactionReplays.tsx
@@ -162,7 +162,7 @@ function ReplaysContent({
     events,
   });
 
-  const {allMobileProj} = useAllMobileProj();
+  const {allMobileProj} = useAllMobileProj({});
 
   return (
     <Layout.Main fullWidth>
