+++ static/app/views/replays/detail/useAllMobileProj.tsx
@@ -1,8 +1,8 @@
-import {mobile} from 'sentry/data/platformCategories';
+import {mobile, replayMobilePlatforms} from 'sentry/data/platformCategories';
 import usePageFilters from 'sentry/utils/usePageFilters';
 import useProjects from 'sentry/utils/useProjects';
 
-export default function useAllMobileProj() {
+export default function useAllMobileProj({replayPlatforms}: {replayPlatforms?: boolean}) {
   const {
     selection: {projects: projectIds},
   } = usePageFilters();
@@ -13,7 +13,9 @@ export default function useAllMobileProj() {
   // if no projects selected, look through all projects
   const proj = projectsSelected.length ? projectsSelected : projects;
 
-  const allMobileProj = proj.every(p => mobile.includes(p.platform ?? 'other'));
+  const allMobileProj = replayPlatforms
+    ? proj.every(p => mobile.includes(p.platform ?? 'other'))
+    : proj.every(p => replayMobilePlatforms.includes(p.platform ?? 'other'));
 
   return {allMobileProj};
 }
